data-explorer/requirements.txt
pandas
langchain
langchain-openai
langchain-anthropic
python-dotenv
